refactor: rename messages to events
